Implemented Comparable
Implemented the comparable interface for the Movie class. It compares the name of the movie and provides a sort by alphabetical order. Changed text of btnCompare to Sort, for clarity, variable names stays the same. Added sort method in MovieList class to use the comparable interface made available in Movie class.